requirements.txt
# ================================================================
# Alpha Trading System — Backend Python 의존성
# 설치: pip install -r requirements.txt --break-system-packages
# tech_stack.md 에 명시된 버전만 사용합니다.
# ================================================================

# ── API 프레임워크 ────────────────────────────────────────────────
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
email-validator>=2.0,<3.0

# ── 데이터베이스 ─────────────────────────────────────────────────
asyncpg>=0.30,<1.0

# ── Redis ────────────────────────────────────────────────────────
redis>=5.0,<6.0

# ── HTTP 클라이언트 ──────────────────────────────────────────────
httpx>=0.27,<1.0

# ── WebSocket (KIS 실시간 틱) ─────────────────────────────────────
websockets>=13,<14

# ── AI 에이전트 오케스트레이션 ────────────────────────────────────
langgraph>=0.2,<1.0

# ── LLM 클라이언트 ──────────────────────────────────────────────
anthropic>=0.40,<1.0
openai>=1.50,<2.0
google-generativeai>=0.8,<1.0
google-auth>=2.0,<3.0

# ── 한국 주식 데이터 ─────────────────────────────────────────────
finance-datareader>=0.9,<1.0
yfinance>=0.2,<2.0

# ── 스케줄링 ─────────────────────────────────────────────────────
apscheduler>=3.10,<4.0

# ── 알림 ─────────────────────────────────────────────────────────
python-telegram-bot>=21,<22

# ── 환경변수 / 유틸 ──────────────────────────────────────────────
python-dotenv>=1.0,<2.0
PyJWT>=2.8,<3.0
